kill process on dropped ping

+++ ps2rpc.py
@@ -119,7 +119,8 @@ def main():
         # check if the last ping was successfull, Clear Rich Presence otherwise
         if not last_ping.value:
             RPC.clear()
-            p.kill()
+            if p.is_alive():
+                p.kill()
         if ip == PS2_IP:
             if not p.is_alive():
                 p.start()
